# Allow reuse of Mapnik objects if enabled