When Mapproxy uses Mapnik without using Renderd (by the way, that's not documented), a new Mapnik map object is initialised for any new rendering request. This harms performance – at least with some map styles – because Mapnik runs all the SQL queries of the layers of the style with an appended `LIMIT 0` to get the types of the returned columns. This pull request allows to reuse map objects.

For example, rendering a map style similar to OSM Carto requires 20 to 30 seconds without, and 6 to 8 seconds with reused Mapnik map objects. Without this patch, Mapproxy performance significantly slower with Mapnik sources compared to a Tirex+mod_tile setup.

Remarks: Multithreading is not supported. If a WSGI worker daemon is restarted, Mapnik is initialised when the first Mapnik rendering is requested.