@ArneBab 2022-03-24
Here’s also a larger (and therefore more risky) pull-request that adds multithreaded operation on mapnik. Maybe it could be useful for you: https://github.com/mapproxy/mapproxy/pull/447/

It uses precise caching and per-thread re-use of mapnik instances.