@weskamm 2023-04-13
superseeded by https://github.com/mapproxy/mapproxy/pull/447